requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.3.0
iniconfig==2.1.0
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.4.0
pytest-django==4.11.1
python-decouple==3.8
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.2
